requirements.txt
asgiref==3.8.1
Django==4.2.13
django-extensions==3.2.3
django-rest==0.8.7
django-soft-delete==1.0.13
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
fpdf==1.7.2
pillow==10.3.0
PyJWT==2.8.0
python-slugify==8.0.4
rest-framework-simplejwt==0.0.2
six==1.16.0
sqlparse==0.5.0
stringify==0.1.1
text-unidecode==1.3
typing_extensions==4.12.1
whitenoise==6.6.0
python-dateutil
PyPDF2